Fix execution of migrations in MySQL
This is a quite empirical fix.
My hypothesis is that the migration squashing algorithm creates
intermediary migration operations which run fine under Postgres
but which create problems with MySQL. In our case, some of the
commented-out operations led to duplicated names for generated
constraints, hence integrity errors while running them.

By "manually squashing" those migration operations, the problems
seem to be avoided.

Keep in mind this is very much a #WorksForUs thing ^^